% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr-sql.R
\name{sql_query_save.PrestoConnection}
\alias{sql_query_save.PrestoConnection}
\title{dbplyr SQL methods}
\usage{
\method{sql_query_save}{PrestoConnection}(con, sql, name, temporary = TRUE, ..., with = NULL)
}
\arguments{
\item{con}{A database connection.}

\item{sql}{a character string containing SQL statement.}

\item{name}{The table name, passed on to \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[=Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[=SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{temporary}{If a temporary table should be created. Default to TRUE in
the \code{\link[dbplyr:db-sql]{dbplyr::sql_query_save()}} generic. The default value generates an
error in Presto. Using \code{temporary = FALSE} to save the query in a
permanent table.}

\item{...}{Other arguments used by individual methods.}

\item{with}{An optional WITH clause for the CREATE TABLE statement.}
}
\description{
dbplyr SQL methods
}
