% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr-db.R
\name{dbplyr_edition.PrestoConnection}
\alias{dbplyr_edition.PrestoConnection}
\title{Inform the dbplyr version used in this package}
\usage{
\method{dbplyr_edition}{PrestoConnection}(con)
}
\arguments{
\item{con}{A DBIConnection object.}
}
\description{
Inform the dbplyr version used in this package
}
