% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRenameTable.R
\name{dbRenameTable}
\alias{dbRenameTable}
\title{Rename a table}
\usage{
dbRenameTable(conn, name, new_name, ...)
}
\arguments{
\item{conn}{A PrestoConnection.}

\item{name}{Existing table's name.}

\item{new_name}{New table name.}

\item{...}{Extra arguments passed to dbExecute.}
}
\description{
Rename a table
}
