% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.data.type.PrestoConnection.R,
%   R/db.explain.PrestoConnection.R, R/db.query.rows.PrestoConnection.R,
%   R/db_desc.PrestoConnection.R, R/db_query_fields.R
\name{db_data_type.PrestoConnection}
\alias{db_data_type.PrestoConnection}
\alias{db_explain.PrestoConnection}
\alias{db_query_rows.PrestoConnection}
\alias{db_desc.PrestoConnection}
\alias{db_query_fields.PrestoConnection}
\title{S3 implementation of \code{\link[dplyr]{db_data_type}} for Presto.}
\usage{
\method{db_data_type}{PrestoConnection}(con, fields, ...)

\method{db_explain}{PrestoConnection}(con, sql, ...)

\method{db_query_rows}{PrestoConnection}(con, sql)

\method{db_desc}{PrestoConnection}(x)

\method{db_query_fields}{PrestoConnection}(con, sql, ...)
}
\description{
S3 implementation of \code{\link[dplyr]{db_data_type}} for Presto.

S3 implementation of \code{\link[dplyr]{db_explain}} for Presto.

S3 implementation of \code{\link[dplyr]{db_query_rows}} for Presto.

S3 implementation of \code{db_desc} for Presto.

S3 implementation of \code{db_query_fields} for Presto.
}
\keyword{internal}
