% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.presto.R
\name{tbl.src_presto}
\alias{tbl.src_presto}
\alias{copy_to.src_presto}
\title{dplyr integration to connect to a table in a database.}
\usage{
\method{tbl}{src_presto}(src, from, ...)

\method{copy_to}{src_presto}(dest, df)
}
\arguments{
\item{src}{A presto src created with \code{src_presto}.}

\item{from}{Either a string giving the name of table in database, or
\code{\link[dplyr]{sql}} described a derived table or compound join.}
}
\description{
Use \code{src_presto} to connect to an existing database,
and \code{tbl} to connect to tables within that database.
If you're unsure of the arguments to pass, please ask your database
administrator for the values of these variables.
}
\examples{
\dontrun{
# First create a database connection with src_presto, then reference a tbl
# within that database
my_tbl <- tbl(my_db, "my_table")
}
}
\keyword{internal}
