% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{dummyPrestoConnection}
\alias{dummyPrestoConnection}
\title{A dummy PrestoConnection}
\usage{
dummyPrestoConnection()
}
\description{
A dummy PrestoConnection
}
\examples{
dummyPrestoConnection()
}
\keyword{internal}
