% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlAppendTableAs.R
\name{get_query_columns}
\alias{get_query_columns}
\title{Get column names from a SQL query or tbl_presto object}
\usage{
get_query_columns(conn, sql)
}
\arguments{
\item{conn}{PrestoConnection object}

\item{sql}{SQL string or tbl_presto object}
}
\value{
Character vector of column names in query order
}
\description{
Get column names from a SQL query or tbl_presto object
}
\keyword{internal}
