% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlAppendTableAs.R
\name{get_table_columns}
\alias{get_table_columns}
\title{Get column names from a table}
\usage{
get_table_columns(conn, name)
}
\arguments{
\item{conn}{PrestoConnection object}

\item{name}{Table name}
}
\value{
Character vector of column names in table order
}
\description{
Get column names from a table
}
\keyword{internal}
