% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{is_simple_table_reference}
\alias{is_simple_table_reference}
\title{Check if a select_query is a simple table reference (no filters, groupings, etc.)}
\usage{
is_simple_table_reference(query)
}
\arguments{
\item{query}{A select_query object}
}
\value{
Logical indicating if the query is a simple table reference
}
\description{
Check if a select_query is a simple table reference (no filters, groupings, etc.)
}
\keyword{internal}
