% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_headers.R
\name{kerberos_configs}
\alias{kerberos_configs}
\title{A convenient wrapper around Kerberos config}
\usage{
kerberos_configs(user = "", password = "", service_name = "presto")
}
\arguments{
\item{user}{User name to pass to httr::authenticate(). Default to "".}

\item{password}{Password to pass to httr::authenticate(). Default to "".}

\item{service_name}{The service name. Default to "presto".}
}
\value{
A httr::config() output that can be passed to the request.config
argument of dbConnect().
}
\description{
The configs specify authentication protocol and additional settings.
}
