% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{new_lazy_unnest_query}
\alias{new_lazy_unnest_query}
\title{Create a lazy_unnest_query object}
\usage{
new_lazy_unnest_query(
  x,
  col,
  values_to = NULL,
  with_ordinality = FALSE,
  ordinality_to = NULL
)
}
\arguments{
\item{x}{Parent lazy_query object}

\item{col}{Column name to unnest (character)}

\item{values_to}{Name of output column for unnested values}

\item{with_ordinality}{Whether to include ordinality column}

\item{ordinality_to}{Name of ordinality column}
}
\description{
Create a lazy_unnest_query object
}
\keyword{internal}
