% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{op_grps.lazy_unnest_query}
\alias{op_grps.lazy_unnest_query}
\title{Get grouping variables for lazy_unnest_query}
\usage{
\method{op_grps}{lazy_unnest_query}(op, ...)
}
\arguments{
\item{op}{lazy_unnest_query object}

\item{...}{Additional arguments}
}
\description{
Get grouping variables for lazy_unnest_query
}
\keyword{internal}
