% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{op_sort.lazy_unnest_query}
\alias{op_sort.lazy_unnest_query}
\title{Get sorting variables for lazy_unnest_query}
\usage{
\method{op_sort}{lazy_unnest_query}(op, ...)
}
\arguments{
\item{op}{lazy_unnest_query object}

\item{...}{Additional arguments}
}
\description{
Get sorting variables for lazy_unnest_query
}
\keyword{internal}
