% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{op_vars.lazy_unnest_query}
\alias{op_vars.lazy_unnest_query}
\title{Get column names for lazy_unnest_query}
\usage{
\method{op_vars}{lazy_unnest_query}(op, ...)
}
\arguments{
\item{op}{lazy_unnest_query object}

\item{...}{Additional arguments}
}
\description{
Get column names for lazy_unnest_query
}
\keyword{internal}
