% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoCursor.R
\docType{class}
\name{PrestoCursor-class}
\alias{PrestoCursor}
\alias{PrestoCursor-class}
\title{Internal implementation detail class needed for its side-effects.
When dbFetch is called, we need to both return the data and update
the uri to the next value.}
\description{
Internal implementation detail class needed for its side-effects.
When dbFetch is called, we need to both return the data and update
the uri to the next value.
}

