% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoDriver.R, R/dbUnloadDriver.R
\docType{class}
\name{PrestoDriver-class}
\alias{PrestoDriver-class}
\alias{show,PrestoDriver-method}
\alias{dbUnloadDriver,PrestoDriver-method}
\title{An S4 class to represent a Presto Driver (and methods)
It is used purely for dispatch and \code{dbUnloadDriver} is unnecessary}
\usage{
\S4method{show}{PrestoDriver}(object)

\S4method{dbUnloadDriver}{PrestoDriver}(drv, ...)
}
\description{
An S4 class to represent a Presto Driver (and methods)
It is used purely for dispatch and \code{dbUnloadDriver} is unnecessary
}
\keyword{internal}
