% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoResult.R, R/dbClearResult.R,
%   R/dbFetch.R, R/dbGetRowCount.R, R/dbGetRowsAffected.R, R/dbGetStatement.R,
%   R/dbHasCompleted.R, R/dbIsValid.R, R/dbListFields.R, R/fetch.R
\docType{class}
\name{PrestoResult-class}
\alias{PrestoResult-class}
\alias{show,PrestoResult-method}
\alias{dbClearResult,PrestoResult-method}
\alias{dbFetch,PrestoResult,numeric-method}
\alias{dbFetch,PrestoResult,missing-method}
\alias{dbGetRowCount,PrestoResult-method}
\alias{dbGetRowsAffected,PrestoResult-method}
\alias{dbGetStatement,PrestoResult-method}
\alias{dbHasCompleted,PrestoResult-method}
\alias{dbIsValid,PrestoResult-method}
\alias{dbListFields,PrestoResult,missing-method}
\alias{fetch,PrestoResult,integer-method}
\alias{fetch,PrestoResult,numeric-method}
\alias{fetch,PrestoResult,missing-method}
\title{An S4 class to represent a Presto Result}
\usage{
\S4method{show}{PrestoResult}(object)

\S4method{dbClearResult}{PrestoResult}(res, ...)

\S4method{dbFetch}{PrestoResult,numeric}(res, n)

\S4method{dbFetch}{PrestoResult,missing}(res)

\S4method{dbGetRowCount}{PrestoResult}(res, ...)

\S4method{dbGetRowsAffected}{PrestoResult}(res)

\S4method{dbGetStatement}{PrestoResult}(res, ...)

\S4method{dbHasCompleted}{PrestoResult}(res, ...)

\S4method{dbIsValid}{PrestoResult}(dbObj, ...)

\S4method{dbListFields}{PrestoResult,missing}(conn, name)

\S4method{fetch}{PrestoResult,integer}(res, n = -1, ...)

\S4method{fetch}{PrestoResult,numeric}(res, n = -1, ...)

\S4method{fetch}{PrestoResult,missing}(res)
}
\description{
An S4 class to represent a Presto Result
}
\section{Slots}{

\describe{
\item{\code{statement}}{The SQL statement sent to the database}

\item{\code{connection}}{The connection object associated with the result}

\item{\code{query}}{An internal implementation detail for keeping track of
what stage a request is in}

\item{\code{post.data}}{Any data extracted from the POST request response}

\item{\code{bigint}}{How bigint type should be handled}
}}

\keyword{internal}
