% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoSession.R
\docType{class}
\name{PrestoSession-class}
\alias{PrestoSession-class}
\alias{PrestoSession}
\title{Class to encapsulate a Presto session}
\description{
Internal implementation detail class needed for its side-effects.
When SET/RESET SESSION queries are called, session parameters need to be
maintained by the client and requires an in-place update.
}
\section{Slots}{

\describe{
\item{\code{.parameters}}{List of Presto session parameters to be added to the
X-Presto-Session header.}
}}


\keyword{internal}
