% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlAppendTableAs.R
\name{reorder_select_columns}
\alias{reorder_select_columns}
\title{Reorder SELECT columns to match target order}
\usage{
reorder_select_columns(conn, sql, target_order)
}
\arguments{
\item{conn}{PrestoConnection object}

\item{sql}{SQL string or tbl_presto object}

\item{target_order}{Character vector of column names in target order}
}
\value{
SQL string or tbl_presto object with reordered columns
}
\description{
Reorder SELECT columns to match target order
}
\keyword{internal}
