% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{sql_build.lazy_unnest_query}
\alias{sql_build.lazy_unnest_query}
\title{Build SQL for lazy_unnest_query}
\usage{
\method{sql_build}{lazy_unnest_query}(op, con, ...)
}
\arguments{
\item{op}{lazy_unnest_query object}

\item{con}{Database connection}

\item{...}{Additional arguments}
}
\description{
Build SQL for lazy_unnest_query
}
\keyword{internal}
