% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{sql_render.lazy_unnest_query}
\alias{sql_render.lazy_unnest_query}
\title{Render SQL for lazy_unnest_query}
\usage{
\method{sql_render}{lazy_unnest_query}(query, con, ...)
}
\arguments{
\item{query}{Query object (either lazy_unnest_query or result of sql_build)}

\item{con}{Database connection (PrestoConnection)}

\item{...}{Additional arguments}
}
\description{
Render SQL for lazy_unnest_query
}
\keyword{internal}
