% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_unnest.R
\name{sql_render.unnest_query}
\alias{sql_render.unnest_query}
\title{Render SQL for unnest_query objects}
\usage{
\method{sql_render}{unnest_query}(query, con, ...)
}
\arguments{
\item{query}{unnest_query object}

\item{con}{Database connection (PrestoConnection)}

\item{...}{Additional arguments}
}
\description{
Render SQL for unnest_query objects
}
\keyword{internal}
