% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlCreateTableAs.R
\name{sqlCreateTableAs}
\alias{sqlCreateTableAs}
\title{Compose query to create a simple table using a statement}
\usage{
sqlCreateTableAs(con, table, statement, with = NULL, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{The table name, passed on to [dbQuoteIdentifier()]. Options are:
- a character string with the unquoted DBMS table name,
  e.g. `"table_name"`,
- a call to [Id()] with components to the fully qualified table name,
  e.g. `Id(schema = "my_schema", table = "table_name")`
- a call to [SQL()] with the quoted and fully qualified table name
  given verbatim, e.g. `SQL('"my_schema"."table_name"')`}

\item{statement}{a character string containing SQL.}

\item{with}{An optional WITH clause for the CREATE TABLE statement.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Compose query to create a simple table using a statement
}
