% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src.presto.R, R/tbl.src.presto.R
\name{src_presto}
\alias{src_presto}
\alias{tbl.src_presto}
\title{dplyr integration to connect to a Presto database.}
\usage{
src_presto(catalog = NULL, schema = NULL, user = NULL, host = NULL,
  port = NULL, parameters = NULL, ...)

tbl.src_presto(src, from, ...)
}
\arguments{
\item{catalog}{Catalog to use in the connection}

\item{schema}{Schema to use in the connection}

\item{user}{User name to use in the connection}

\item{host}{Host name to connect to the database}

\item{port}{Port number to use with the host name}

\item{parameters}{Additional parameters to pass to the connection}

\item{...}{Other arguments passed on to the underlying
database connector \code{dbConnect}}

\item{src}{A presto src created with \code{src_presto}.}

\item{from}{Either a string giving the name of table in database, or
\code{\link{sql}} described a derived table or compound join.}

\item{...}{Included for compatibility with the generic, but otherwise
ignored.}
}
\description{
Allows you to connect to an existing database through a presto connection.

Use \code{src_presto} to connect to an existing database,
and \code{tbl} to connect to tables within that database.
If you're unsure of the arguments to pass, please ask your database
administrator for the values of these variables.
}
\examples{
\dontrun{
# To connect to a database
my_db <- src_presto(catalog = "hive", schema = "web", user = "onur",
  host = "localhost", port = 8888)
}
\dontrun{
First create a database connection with src_presto, then reference a tbl
within that database
my_tbl <- tbl(my_db, "my_table")
}
}

