% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlAppendTableAs.R
\name{validate_column_match}
\alias{validate_column_match}
\title{Validate that column names match between table and query}
\usage{
validate_column_match(table_cols, query_cols)
}
\arguments{
\item{table_cols}{Character vector of table column names}

\item{query_cols}{Character vector of query column names}
}
\value{
Invisibly returns TRUE if match, otherwise throws error
}
\description{
Validate that column names match between table and query
}
\keyword{internal}
