% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.data.type.PrestoConnection.R,
%   R/db.explain.PrestoConnection.R, R/db.query.rows.PrestoConnection.R,
%   R/db_collect.R, R/db_desc.PrestoConnection.R, R/db_query_fields.R,
%   R/tbl.src.presto.R
\name{db_data_type.PrestoConnection}
\alias{db_data_type.PrestoConnection}
\alias{db_explain.PrestoConnection}
\alias{db_query_rows.PrestoConnection}
\alias{db_collect.PrestoConnection}
\alias{db_desc.PrestoConnection}
\alias{db_query_fields.PrestoConnection}
\alias{collect.tbl_presto}
\title{S3 implementation of \code{\link[dplyr]{db_data_type}} for Presto.}
\usage{
\method{db_data_type}{PrestoConnection}(con, fields, ...)

\method{db_explain}{PrestoConnection}(con, sql, ...)

\method{db_query_rows}{PrestoConnection}(con, sql)

\method{db_collect}{PrestoConnection}(con, sql, n = -1, warn_incomplete = TRUE, ...)

\method{db_desc}{PrestoConnection}(x)

\method{db_query_fields}{PrestoConnection}(con, sql, ...)

\method{collect}{tbl_presto}(x, ..., n = Inf, warn_incomplete = TRUE)
}
\description{
S3 implementation of \code{\link[dplyr]{db_data_type}} for Presto.

S3 implementation of \code{\link[dplyr]{db_explain}} for Presto.

S3 implementation of \code{\link[dplyr]{db_query_rows}} for Presto.

S3 implementation of \code{db_collect} for Presto.

S3 implementation of \code{db_desc} for Presto.

S3 implementation of \code{db_query_fields} for Presto.

S3 implementation of \code{collect} for Presto.
}
\keyword{internal}
