% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.to.src.presto.R,
%   R/db.data.type.PrestoConnection.R, R/db.explain.PrestoConnection.R,
%   R/db.query.rows.PrestoConnection.R, R/db_desc.PrestoConnection.R,
%   R/db_query_fields.R, R/sql_translate_env.R
\name{copy_to.src_presto}
\alias{copy_to.src_presto}
\alias{db_data_type.PrestoConnection}
\alias{db_explain.PrestoConnection}
\alias{db_query_rows.PrestoConnection}
\alias{db_desc.PrestoConnection}
\alias{db_query_fields.PrestoConnection}
\alias{sql_translate_env.PrestoConnection}
\title{S3 implementation of \code{\link[dplyr]{copy_to}} for Presto.}
\usage{
copy_to.src_presto(dest, df)

db_data_type.PrestoConnection(con, fields, ...)

db_explain.PrestoConnection(con, sql, ...)

db_query_rows.PrestoConnection(con, sql)

db_desc.PrestoConnection(x)

db_query_fields.PrestoConnection(con, sql, ...)

sql_translate_env.PrestoConnection(con)
}
\description{
S3 implementation of \code{\link[dplyr]{copy_to}} for Presto.

S3 implementation of \code{\link[dplyr]{db_data_type}} for Presto.

S3 implementation of \code{\link[dplyr]{db_explain}} for Presto.

S3 implementation of \code{\link[dplyr]{db_query_rows}} for Presto.

S3 implementation of \code{db_desc} for Presto.

S3 implementation of \code{db_query_fields} for Presto.

S3 implementation of \code{sql_translate_env} for Presto.
}
\keyword{internal}
