% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoSession.R
\docType{class}
\name{PrestoSession-class}
\alias{PrestoSession-class}
\alias{PrestoSession}
\title{Internal implementation detail class needed for its side-effects.
When SET/RESET SESSION queries are called, session parameters need to be
maintained by the client and requires an in-place update.}
\description{
Internal implementation detail class needed for its side-effects.
When SET/RESET SESSION queries are called, session parameters need to be
maintained by the client and requires an in-place update.
}

