% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPresto-package.R
\docType{package}
\name{RPresto-package}
\alias{RPresto}
\alias{RPresto-package}
\title{RPresto: DBI Connector to Presto}
\description{
Implements a 'DBI' compliant interface to Presto. Presto is an open source distributed SQL query engine for running interactive analytic queries against data sources of all sizes ranging from gigabytes to petabytes: \url{https://prestodb.io/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prestodb/RPresto}
  \item Report bugs at \url{https://github.com/prestodb/RPresto/issues}
}

}
\author{
\strong{Maintainer}: Thomas J. Leeper \email{thomasleeper@fb.com} [contributor]

Authors:
\itemize{
  \item Onur Ismail Filiz \email{onur@fb.com}
  \item Sergey Goder \email{sgoder@fb.com}
}

Other contributors:
\itemize{
  \item John Myles White \email{johnmyleswhite@fb.com} [contributor]
  \item Jarod G.R. Meng \email{jarodm@fb.com} [contributor]
}

}
\keyword{internal}
