% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPresto-package.R
\docType{package}
\name{RPresto-package}
\alias{RPresto}
\alias{RPresto-package}
\title{RPresto: DBI Connector to Presto}
\description{
Implements a 'DBI' compliant interface to Presto. Presto is an open source distributed SQL query engine for running interactive analytic queries against data sources of all sizes ranging from gigabytes to petabytes: \url{https://prestodb.io/}.
}
\section{Package options}{

\describe{
\item{\code{rpresto.max.rows}}{Max number of rows to be fetched before
a warning is given.}

\item{\code{rpresto.quiet}}{Verbose output during processing? The default
value, \code{NA}, turns on verbose output for interactive queries that
run longer than two seconds.  Use \code{FALSE} for immediate verbose
output, \code{TRUE} for quiet operation.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prestodb/RPresto}
  \item Report bugs at \url{https://github.com/prestodb/RPresto/issues}
}

}
\author{
\strong{Maintainer}: Jarod G.R. Meng \email{jarodm@fb.com}

Authors:
\itemize{
  \item Onur Ismail Filiz \email{onur@fb.com}
  \item Sergey Goder \email{sgoder@fb.com}
}

Other contributors:
\itemize{
  \item Thomas J. Leeper \email{thomasleeper@fb.com} [contributor]
  \item John Myles White \email{johnmyleswhite@fb.com} [contributor]
}

}
\keyword{internal}
