% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlCreateTableAs.R
\name{sqlCreateTableAs}
\alias{sqlCreateTableAs}
\title{Compose query to create a simple table using a statement}
\usage{
sqlCreateTableAs(con, name, sql, with = NULL, ...)
}
\arguments{
\item{con}{A database connection.}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{DBI::dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{DBI::Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{DBI::SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{sql}{a character string containing SQL statement.}

\item{with}{An optional WITH clause for the CREATE TABLE statement.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Compose query to create a simple table using a statement
}
