% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.presto.R
\name{src_presto}
\alias{src_presto}
\title{dplyr integration to connect to a Presto database.}
\usage{
src_presto(
  catalog = NULL,
  schema = NULL,
  user = NULL,
  host = NULL,
  port = NULL,
  source = NULL,
  session.timezone = NULL,
  parameters = NULL,
  ...
)
}
\arguments{
\item{catalog}{Catalog to use in the connection}

\item{schema}{Schema to use in the connection}

\item{user}{User name to use in the connection}

\item{host}{Host name to connect to the database}

\item{port}{Port number to use with the host name}

\item{source}{Source to specify for the connection}

\item{session.timezone}{Time zone for the connection}

\item{parameters}{Additional parameters to pass to the connection}

\item{...}{For \code{src_presto} other arguments passed on to the underlying
database connector \code{dbConnect}. For \code{tbl.src_presto}, it is
included for compatibility with the generic, but otherwise ignored.}
}
\description{
Allows you to connect to an existing database through a presto connection.
}
\examples{
\dontrun{
# To connect to a database
my_db <- src_presto(
  catalog = "memory",
  schema = "default",
  user = Sys.getenv("USER"),
  host = "http://localhost",
  port = 8080,
  session.timezone = "Asia/Kathmandu"
)
}
}
